[FIX] looped import
